# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Film API Documentation",
        "description": "API documentation for Film management",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/kinos": {
            "get": {
                "tags": [
                    "Kinos"
                ],
                "summary": "Liste aller Kinos",
                "operationId": "29bb2fd11976b4b0fd41d02f9422d40a",
                "responses": {
                    "200": {
                        "description": "Liste von Kinos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Kino"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Kinos"
                ],
                "summary": "Erstelle ein neues Kino",
                "operationId": "1c430015a7760f0f49d265c97f6b4fc8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/KinoInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kino erfolgreich erstellt",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Kino"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/kinos/{id}": {
            "get": {
                "tags": [
                    "Kinos"
                ],
                "summary": "Zeige ein Kino",
                "operationId": "f3db480a7148a688e72ea34c36a57b07",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID des Kinos",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kino",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Kino"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kino nicht gefunden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kino nicht gefunden."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "resource_not_found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Kinos"
                ],
                "summary": "Aktualisiere ein Kino",
                "operationId": "fd9e797ab9983715ce4ccb7ea7eb6906",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID des Kinos",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/KinoInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kino aktualisiert",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Kino"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kino nicht gefunden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kino nicht gefunden."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "resource_not_found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Kinos"
                ],
                "summary": "Lösche ein Kino",
                "operationId": "e92eb530f9f775c38a40f2da76c89d43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID des Kinos",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Kino gelöscht"
                    },
                    "404": {
                        "description": "Kino nicht gefunden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kino nicht gefunden."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "resource_not_found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Kino wird noch verwendet",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kino kann nicht gelöscht werden, da es von einem oder mehreren Sälen verwendet wird."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "foreign_key_constraint_violation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Kino": {
                "required": [
                    "id",
                    "name",
                    "ort"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Cinépolis"
                    },
                    "ort": {
                        "type": "string",
                        "example": "Zürich"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "KinoInput": {
                "required": [
                    "name",
                    "ort"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Cinépolis"
                    },
                    "ort": {
                        "type": "string",
                        "example": "Zürich"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}